FEATURE - EC13 chamber control/monitoring system
The EC13 chamber now can work with both usb gpib controllers, the NI one
as well as the Prologix. The later is prefered, since it works on linux
and is almost an order of magnitude cheaper.
The test program "risingsun" for now monitores the temperature, scripts
are still needed for the control part, however some functionality is
there, encapsulated in a "SunChamber" class.