REFACTOR - completed ongoing work, API ready
Rationale: Digitizer is now an object and can be intialized and
configured easily.

FIX - CRITICAL fix, the previous version of the software DID NOT
NECESSARILY RECORD ALL the events, this *should* be fixed now. The
issu was with a pointer to an array, which points only to the first
entry, instead of the array [which is a feature, but ><].