FEATURE - restructures the constructors
This give outside access to the handle via get_handle. Having a handle
allows for multiple connection through the functional style CAEN
library.

The way how this is achieved here is by opening a digitizer with the no
argument constructor, then set the handle via set_handle. This allows
for quering the temperatures and so on.

This closes issue #2