FEATURES - memory management improvements, basline guessing
Rationale: The memory management is a mess, and we encountered
some bad segfaults, which seem to occur in same cases when
ReadEvents is called.
FreeEvent calls have been added, as well as a reset_memory function
has been implemented, which is a shortcut to reallocate memory.

The expected baseline is the bare expectation of the baseline from
the set dc offset.